requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
# Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.10.1
